Replace pelite with custom PE and resource parsing code
This was quite a heavy-weight dependency (due to proc macros?), and we were
barely using any of it.

+++ crates/unzbd/src/main.rs
@@ -58,7 +58,7 @@ struct MsgOpts {
     #[clap(long = "dump-ids", help = "Dump message IDs")]
     dump_ids: bool,
     #[clap(long = "skip-data", hide = true)]
-    skip_data: Option<u64>,
+    skip_data: Option<usize>,
 }
 
 #[derive(Parser)]
